Code improvement - THO-197

+++ SobrassadaEngine/UI/EngineEditors/Editor/StateMachineEditor.cpp
@@ -1,8 +1,9 @@
 #include "StateMachineEditor.h"
-#include "StateNode.h"
-#include "FileSystem/StateMachineManager.h"
+
 #include "Application.h"
+#include "FileSystem/StateMachineManager.h"
 #include "ResourcesModule.h"
+#include "StateNode.h"
 
 StateMachineEditor::StateMachineEditor(const std::string& editorName, UID uid, ResourceStateMachine* stateMachine)
     : EngineEditorBase(editorName, uid), uid(uid), resource(stateMachine)
@@ -20,7 +21,7 @@ bool StateMachineEditor::RenderEditor()
 
     ImGui::Begin(name.c_str());
 
-     if (ImGui::Button("Save"))
+    if (ImGui::Button("Save"))
     {
         SaveMachine();
     }
@@ -41,7 +42,6 @@ bool StateMachineEditor::RenderEditor()
             selectedNode = dynamic_cast<StateNode*>(node.get());
         }
     }
-   
 
     graph->rightClickPopUpContent(
         [this](ImFlow::BaseNode* node)
@@ -56,10 +56,11 @@ bool StateMachineEditor::RenderEditor()
                         RemoveStateNode(*stateNode);
                     }
                 }
-            }else if (node == nullptr && ImGui::MenuItem("Add State"))
+            }
+            else if (node == nullptr && ImGui::MenuItem("Add State"))
             {
-                ImVec2 pos = graph->screen2grid(ImGui::GetMousePos());
-            
+                ImVec2 pos   = graph->screen2grid(ImGui::GetMousePos());
+
                 auto newNode = graph->placeNodeAt<StateNode>(pos);
                 if (newNode)
                 {
@@ -74,7 +75,7 @@ bool StateMachineEditor::RenderEditor()
     );
 
     DetectNewTransitions();
-    
+
     graph->droppedLinkPopUpContent(
         [this](ImFlow::Pin* dragged)
         {
@@ -105,15 +106,14 @@ bool StateMachineEditor::RenderEditor()
 
     ImGui::End();
 
-    //INSPECTOR
+    // INSPECTOR
     if (selectedNode != nullptr)
     {
         if (ImGui::IsKeyPressed(ImGuiKey_Delete) && !selectedNode->toDestroy())
         {
-            DeleteStateResource(*selectedNode); 
-            selectedNode->destroy();            
+            DeleteStateResource(*selectedNode);
+            selectedNode->destroy();
         }
-
 
         if (!selectedNode->toDestroy())
         {
@@ -137,10 +137,9 @@ void StateMachineEditor::ShowInspector()
     ImGui::Separator();
 
     const std::string& stateName = selectedNode->GetStateName();
-    char nameBuffer[128]  = "";
+    char nameBuffer[128]         = "";
     strncpy_s(nameBuffer, stateName.c_str(), sizeof(nameBuffer));
     nameBuffer[sizeof(nameBuffer) - 1] = '\0';
-
 
     if (ImGui::InputText("State Name", nameBuffer, sizeof(nameBuffer)))
     {
@@ -207,7 +206,7 @@ void StateMachineEditor::ShowInspector()
                 if (out_text) *out_text = (*clips)[idx].c_str();
                 return true;
             },
-            &availableClips, availableClips.size()
+            &availableClips, (int)availableClips.size()
         ))
     {
         if (currentClipIndex >= 0 && availableClips[currentClipIndex] != selectedNode->GetClipName())
@@ -294,7 +293,7 @@ void StateMachineEditor::ShowInspector()
                         if (out_text) *out_text = (*triggers)[idx].c_str();
                         return true;
                     },
-                    &availableTriggers, availableTriggers.size()
+                    &availableTriggers, (int)availableTriggers.size()
                 ))
             {
                 if (currentTriggerIndex >= 0 &&
@@ -349,7 +348,7 @@ void StateMachineEditor::ShowInspector()
                         if (out_text) *out_text = (*triggers)[idx].c_str();
                         return true;
                     },
-                    &availableTriggers, availableTriggers.size()
+                    &availableTriggers, (int)availableTriggers.size()
                 ))
             {
                 if (currentTriggerIndex >= 0 &&
@@ -374,8 +373,6 @@ void StateMachineEditor::ShowInspector()
     }
 
     ImGui::End();
-
-
 }
 
 void StateMachineEditor::BuildGraph()
@@ -392,14 +389,13 @@ void StateMachineEditor::BuildGraph()
                     pin->deleteLink();
                 }
 
-                
                 for (const auto& pin : node->getOuts())
                 {
                     pin->deleteLink();
                 }
             }
         }
-        graph->getNodes().clear(); 
+        graph->getNodes().clear();
     }
     graph = std::make_unique<ImFlow::ImNodeFlow>("StateMachineGraph_" + std::to_string(uid));
 
@@ -408,7 +404,7 @@ void StateMachineEditor::BuildGraph()
 
     for (const auto& state : resource->states)
     {
-        ImVec2 position = state.position; 
+        ImVec2 position = state.position;
 
         auto newNode    = graph->placeNodeAt<StateNode>(position);
         if (newNode)
@@ -422,7 +418,6 @@ void StateMachineEditor::BuildGraph()
 
                 stateNodes[state.name.GetString()] = stateNode;
             }
-            
         }
         stateCont++;
     }
@@ -464,7 +459,7 @@ void StateMachineEditor::DetectNewTransitions()
     {
         if (const auto link = linkWeak.lock())
         {
-            ImFlow::Pin* startPin = link->left(); 
+            ImFlow::Pin* startPin = link->left();
             ImFlow::Pin* endPin   = link->right();
 
             if (!startPin || !endPin) continue;
@@ -519,10 +514,10 @@ void StateMachineEditor::ShowSavePopup()
         {
             allStateMachineNames = StateMachineManager::GetAllStateMachineNames();
 
-            //Save all node positions
+            // Save all node positions
             for (auto& state : resource->states)
             {
-                for (const auto& [uid, node] : graph->getNodes()) 
+                for (const auto& [uid, node] : graph->getNodes())
                 {
                     if (node->getName() == state.name.GetString())
                     {
@@ -557,7 +552,7 @@ void StateMachineEditor::ShowSavePopup()
                 resource->SetName(std::string(stateMachineName));
                 StateMachineManager::SaveStateMachine(resource, alreadySaved);
                 ImGui::CloseCurrentPopup();
-                saveInitialized = false; 
+                saveInitialized = false;
             }
         }
 
@@ -565,7 +560,7 @@ void StateMachineEditor::ShowSavePopup()
         if (ImGui::Button("Cancel"))
         {
             ImGui::CloseCurrentPopup();
-            saveInitialized = false; 
+            saveInitialized = false;
         }
         ImGui::EndPopup();
     }
@@ -581,7 +576,8 @@ void StateMachineEditor::LoadMachine()
 void StateMachineEditor::ShowLoadPopup()
 {
 
-    if (ImGui::BeginPopupModal("Load State Machine", NULL, ImGuiWindowFlags_AlwaysAutoResize)){
+    if (ImGui::BeginPopupModal("Load State Machine", NULL, ImGuiWindowFlags_AlwaysAutoResize))
+    {
         if (allStateMachineNames.empty())
         {
             allStateMachineNames = StateMachineManager::GetAllStateMachineNames();
@@ -610,7 +606,8 @@ void StateMachineEditor::ShowLoadPopup()
             {
                 GLOG("Loading State Machine: %s", selectedName.c_str());
 
-                ResourceStateMachine* stateMachine = (ResourceStateMachine*)App->GetResourcesModule()->RequestResource(selectedUID);
+                ResourceStateMachine* stateMachine =
+                    (ResourceStateMachine*)App->GetResourcesModule()->RequestResource(selectedUID);
                 if (stateMachine)
                 {
                     GLOG("Successfully loaded State Machine: %s", selectedName.c_str());
@@ -623,7 +620,6 @@ void StateMachineEditor::ShowLoadPopup()
                     GLOG("Failed to load State Machine: %s", selectedName.c_str());
                 }
             }
-        
         }
         ImGui::SameLine();
 
@@ -634,7 +630,6 @@ void StateMachineEditor::ShowLoadPopup()
 
         ImGui::EndPopup();
     }
-
 }
 void StateMachineEditor::RemoveStateNode(StateNode& node)
 {
@@ -645,11 +640,10 @@ void StateMachineEditor::RemoveStateNode(StateNode& node)
     auto it     = nodes.find(node.getUID());
     if (it != nodes.end())
     {
-        nodes.erase(it); 
+        nodes.erase(it);
     }
 
     selectedNode = nullptr;
-
 }
 
 void StateMachineEditor::DeleteStateResource(StateNode& node)
@@ -666,7 +660,7 @@ void StateMachineEditor::DeleteStateResource(StateNode& node)
     }
 
     const State* state = resource->GetState(stateName);
-    clipName   = state->clipName.GetString();
+    clipName           = state->clipName.GetString();
     resource->RemoveState(state->name.GetString());
 
     const Clip* clip = resource->GetClip(clipName);
